refactor: rename CharacterFilterDataUpdate event
